Chore: move constants to constants.js file + add requests file for api config
