+++ src/containers/HomeTemplate/components/Footer/index.jsx
@@ -1,161 +1,16 @@
-import { Button, Grid, IconButton, List, ListItem, Menu, MenuItem } from "@mui/material";
-import LanguageIcon from "@mui/icons-material/Language";
-import EuroIcon from "@mui/icons-material/Euro";
-import { Container } from "@mui/system";
 import React from "react";
 import { Link } from "react-router-dom";
 
+//Material UI
+import { Grid, IconButton, List, ListItem } from "@mui/material";
+import { Container } from "@mui/system";
+import { Circle } from "@mui/icons-material";
+
 //others
 import "./style.scss";
-import { Circle, Facebook, Instagram, Twitter } from "@mui/icons-material";
+import { footerCategories, footerRightList, userTools, socialList } from "./constants";
 
 function Footer() {
-    const footerCategories = [
-        {
-            categoryName: "Support",
-            categoryItem: [
-                {
-                    name: "Help Center",
-                    link: "/",
-                },
-                {
-                    name: "AirCover",
-                    link: "/",
-                },
-                {
-                    name: "Safety information",
-                    link: "/",
-                },
-                {
-                    name: "Supporting people with disabilities",
-                    link: "/",
-                },
-                {
-                    name: "Cancellation options",
-                    link: "/",
-                },
-                {
-                    name: "Our COVID-19 Response",
-                    link: "/",
-                },
-                {
-                    name: "Report a neighborhood concern",
-                    link: "/",
-                },
-            ],
-        },
-        {
-            categoryName: "Community",
-            categoryItem: [
-                {
-                    name: "Airbnb.org: disaster relief housing",
-                    link: "/",
-                },
-                {
-                    name: "Support Afghan refugees",
-                    link: "/",
-                },
-                {
-                    name: "Combating discrimination",
-                    link: "/",
-                },
-            ],
-        },
-        {
-            categoryName: "Hosting",
-            categoryItem: [
-                {
-                    name: "Try hosting",
-                    link: "/",
-                },
-                {
-                    name: "AirCover for Hosts",
-                    link: "/",
-                },
-                {
-                    name: "Explore hosting resources",
-                    link: "/",
-                },
-                {
-                    name: "Visit our community forum",
-                    link: "/",
-                },
-                {
-                    name: "How to host responsibly",
-                    link: "/",
-                },
-            ],
-        },
-        {
-            categoryName: "Airbnb",
-            categoryItem: [
-                {
-                    name: "Newsroom",
-                    link: "/",
-                },
-                {
-                    name: "Learn about new features",
-                    link: "/",
-                },
-                {
-                    name: "Letter from our founders",
-                    link: "/",
-                },
-                {
-                    name: "Investors",
-                    link: "/",
-                },
-                {
-                    name: "Gift cards",
-                    link: "/",
-                },
-            ],
-        },
-    ];
-
-    const footerRightList = [
-        {
-            name: "Privacy",
-            link: "/",
-        },
-        {
-            name: "Terms",
-            link: "/",
-        },
-        {
-            name: "Sitemap",
-            link: "/",
-        },
-    ];
-
-    const userTools = [
-        {
-            name: "Language",
-            icon: <LanguageIcon className="main-footer__tool-icon" />,
-            link: "/",
-        },
-        {
-            name: "Currency",
-            icon: <EuroIcon className="main-footer__tool-icon" />,
-            link: "/",
-        },
-    ];
-
-    const socialList = [
-        {
-            icon: <Facebook className="main-footer__social-item-icon" />,
-            link: "/",
-        },
-        {
-            icon: <Twitter className="main-footer__social-item-icon" />,
-            link: "/",
-        },
-        {
-            icon: <Instagram className="main-footer__social-item-icon" />,
-            link: "/",
-        },
-    ];
-
     const renderFooterCategories = () => {
         return footerCategories.map((category, index) => (
             <Grid item xs={12} sm={6} md={3} key={index} className="main-footer__category">
@@ -190,7 +45,7 @@ function Footer() {
 
     const renderUserTools = () => {
         return userTools.map((item, index) => (
-            <Grid item xs={4} sm={4} key={index} className="main-footer__tool-item" href={item.link}>
+            <Grid item xs={4} sm={5} key={index} className="main-footer__tool-item" href={item.link}>
                 {item.icon}
                 <span className="main-footer__tool-name">{item.name}</span>
             </Grid>
