+++ src/containers/HomeTemplate/components/Header/constants.js
@@ -1,0 +1,30 @@
+const pages = ["Location", "Stays", "Online Experiences"];
+const settings = [
+    {
+        label: "Sign up",
+        link: "/auth/login",
+        divider: false,
+    },
+    {
+        label: "Log in",
+        link: "/auth/signup",
+        divider: true,
+    },
+    {
+        label: "Host your home",
+        link: "/",
+        divider: false,
+    },
+    {
+        label: "Host an experience",
+        link: "/",
+        divider: false,
+    },
+    {
+        label: "Help",
+        link: "/",
+        divider: false,
+    },
+];
+
+export { pages, settings };
