+++ src/containers/HomeTemplate/components/RoomList/constants.js
@@ -1,0 +1,22 @@
+//Components
+import images from "@/assets/images";
+
+const roomCategories = [
+    {
+        name: "Entire place",
+        alt: "entire place",
+        image: images.roomListEntire,
+    },
+    {
+        name: "Modern place",
+        alt: "modern place",
+        image: images.roomListModern,
+    },
+    {
+        name: "Surrounded by Nature",
+        alt: "places surrounded by Nature",
+        image: images.roomListNatural,
+    },
+];
+
+export { roomCategories };
