+++ src/containers/HomeTemplate/components/SearchBar/style.scss
@@ -32,12 +32,6 @@ div.search-bar__input-control {
     width: 100%;
 }
 
-// div.search-bar__input-control.active,
-// div.search-bar__input-control.search-bar__input-control--where.active {
-//     background-color: var(--white);
-//     box-shadow: 0 6px 20px rgba(0, 0, 0, 0.2);
-// }
-
 .search-bar__input {
     .MuiOutlinedInput-notchedOutline {
         border: none;
@@ -127,9 +121,6 @@ div.search-bar__input-control {
         &:last-of-type {
             flex: 1.2;
         }
-    }
-
-    div.search-bar__input-control {
     }
 }
